requirements.txt
Django==3.2.15
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
python-dotenv==0.21.0
django-filter==22.1
psycopg2-binary==2.9.3
django-cors-headers==3.13.0
celery==5.2.7
django-celery-beat==2.4.0
django_rest_swagger==2.2.0